# Conflicts:
#	WiseOldManConnector/Models/Output/Delta.cs

+++ WiseOldManConnector/Models/API/Responses/WOMGroupDeltaMember.cs
@@ -10,7 +10,7 @@ namespace WiseOldManConnector.Models.API.Responses {
         public DateTime EndDate { get; set; }
 
         [JsonProperty("gained")]
-        public int Gained { get; set; }
+        public double Gained { get; set; }
 
         [JsonProperty("player")]
         public PlayerResponse Player { get; set; }
